refactor: port to typescript
