--- src/index.d.ts
@@ -1,6 +1,0 @@
-export function useContextualRouting(): {
-  returnHref: string;
-  makeContextualHref: (extraQueryParams: { [key: string]: any }) => string;
-};
-
-
